Refactor: Extract test running into a dedicated class
