+++ src/main/php/unittest/TestInstance.class.php
@@ -27,11 +27,12 @@ class TestInstance extends TestGroup {
     $this->instance= $instance;
   }
 
+  /** @return lang.XPClass */
+  public function type() { return typeof($this->instance); }
+
   /** @return int */
   public function numTests() { return 1; }
 
-  /** @return php.Generator */
-  public function tests() {
-    yield $this->instance;
-  }
+  /** @return iterable */
+  public function tests() { yield $this->instance; }
 }
